requirements.txt
flask~=3.0.2
pyodbc~=5.1.0
mysql-connector-python~=8.3.0
Jinja2~=3.1.3
waitress~=3.0.0
PyQt5~=5.15.10
pip~=22.3.1
wheel~=0.38.4
pillow~=10.2.0
MarkupSafe~=2.1.5
werkzeug~=3.0.1
click~=8.1.7
blinker~=1.7.0
itsdangerous~=2.1.2
setuptools~=65.5.1
altgraph~=0.17.4
pefile~=2023.2.7
packaging~=23.2
pyinstaller~=6.4.0
fpdf~=1.7.2
pdf2image~=1.17.0